+++ src/main/java/jm/audio/synth/Comb.java
@@ -92,7 +92,7 @@ public final class Comb extends AudioObject {
    *
    */
   public void build() {
-    int sampleDelay = (int) ((float) ((float) this.delay / (float) 1000) * this.sampleRate);
+    int sampleDelay = (int) ((float) this.delay / (float) 1000 * this.sampleRate);
     this.delayLine = new float[sampleDelay * this.channels];
     this.delayIndex = 0;
   }
