+++ src/main/java/jm/audio/synth/Pluck.java
@@ -137,7 +137,7 @@ public final class Pluck extends AudioObject {
       }
     }
     // setup all pass values
-    int sampleDelay = (int) ((float) ((float) this.delay / (float) 1000) * this.sampleRate);
+    int sampleDelay = (int) ((float) this.delay / (float) 1000 * this.sampleRate);
     this.delayLine = new float[sampleDelay * this.channels];
     this.delayIndex = 0;
   }
