+++ src/main/java/jm/gui/sketch/SketchRuler.java
@@ -124,7 +124,7 @@ public class SketchRuler extends Canvas implements MouseListener, MouseMotionLis
   public void mouseDragged(MouseEvent e) {
     //System.out.println("Dragged");
     double beatWidth = sketchScore.beatWidth;
-    beatWidth += (double) ((double) e.getX() - (double) startX) / 5.0;
+    beatWidth += ((double) e.getX() - (double) startX) / 5.0;
     if (beatWidth < 1.0) {
       beatWidth = 1.0;
     }
